+++ .eslintrc.js
@@ -1,10 +1,12 @@
+'use strict';
+
 module.exports = {
-  extends: ["ash-nazg/sauron"],
+  extends: ['ash-nazg/sauron'],
   parserOptions: {
-    sourceType: "module"
+    sourceType: 'module'
   },
   env: {
-    node: true,
+    node: true
   },
   settings: {
     polyfills: [
@@ -13,9 +15,18 @@ module.exports = {
   },
   overrides: [
     {
+      files: '.eslintrc.js',
+      extends: ['plugin:node/recommended-script'],
+      rules: {
+        'import/no-commonjs': 0
+      }
+    },
+    {
       files: ['*.md'],
       rules: {
-        'import/no-unresolved': ['error', { ignore: ['i18nizeelement'] }]
+        'import/no-unresolved': ['error', {
+          ignore: ['i18nizeelement']
+        }]
       }
     },
     {