+++ test/bootstrap/node.js
@@ -1,5 +1,4 @@
 import jsdom from 'jsdom';
-import {expect as chaiExpect} from 'chai';
 import jml from 'jamilih';
 
 const {JSDOM} = jsdom;
@@ -18,5 +17,4 @@ global.window.Element.prototype.closest = function (selector) {
   }
   return null;
 };
-global.expect = chaiExpect;
 global.jml = jml;