+++ CHANGES.md
@@ -4,6 +4,11 @@
 
 - Docs: Update Jamilih API; document additional Jamilih plugin APIs
 - Linting (ESLint): As per latest ash-nazg
+- Testing: Use chai/register-expect
+- npm: Add eslint to test script
+- npm: Script from `prepare` to `prepublishOnly`
+- npm: Update `rollup-plugin-babel` to `@rollup/plugin-babel`
+    and make explicit `babelHelpers` value of `bundled`
 - npm: Switch to non-deprecated `@rollup/plugin-node-resolve`
 - npm: Update devDeps